net: add lldap port
Summary:
The library ldap3(dependency) use patch.crates-io section to overwrite one of
the crate's location. However, we use patch section also to change the
remote crate to local store one. We cannot patch a patch in cargo.
There is a workaround now by clone the repo manually and modify the
patch section in the Cargo.toml to our manually clone one.

Differential Revision: https://reviews.freebsd.org/D49986